+++ backend/src/main/java/br/com/gestorfinanceiro/controller/AuthController.java
@@ -1,11 +1,11 @@
 package br.com.gestorfinanceiro.controller;
 
+import br.com.gestorfinanceiro.config.security.JwtUtil;
 import br.com.gestorfinanceiro.dto.LoginDTO;
 import br.com.gestorfinanceiro.dto.UserDTO;
 import br.com.gestorfinanceiro.mappers.Mapper;
 import br.com.gestorfinanceiro.models.UserEntity;
 import br.com.gestorfinanceiro.services.AuthService;
-import br.com.gestorfinanceiro.services.JwtUtil;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
