+++ backend/src/main/java/br/com/gestorfinanceiro/controller/ReceitaController.java
@@ -1,11 +1,11 @@
 package br.com.gestorfinanceiro.controller;
 
+import br.com.gestorfinanceiro.config.security.JwtUtil;
 import br.com.gestorfinanceiro.dto.GraficoBarraDTO;
 import br.com.gestorfinanceiro.dto.GraficoPizzaDTO;
 import br.com.gestorfinanceiro.dto.ReceitaDTO;
 import br.com.gestorfinanceiro.mappers.Mapper;
 import br.com.gestorfinanceiro.models.ReceitaEntity;
-import br.com.gestorfinanceiro.services.JwtUtil;
 import br.com.gestorfinanceiro.services.ReceitaService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
